adjusts p5 drawings after R255 print test

+++ 0x0A-Perlin/p5_0x0A_Perlin/output.pde
@@ -5,17 +5,17 @@ void drawOutput(PGraphics mpg) {
   mpg.beginDraw();
   mpg.rectMode(CENTER);
 
-  for (float x = 0; x < mpg.width + PWIDTH; x += PWIDTH) {
-    for (float y = 0; y < mpg.height + PWIDTH; y += PWIDTH) {
+  for (int x = 0; x < mpg.width + PWIDTH; x += PWIDTH) {
+    for (int y = 0; y < mpg.height + PWIDTH; y += PWIDTH) {
       int alpha = int(255f * mPerlin.noise(x * NOISE_SCALE, y * NOISE_SCALE));
 
       mpg.stroke(COLOR_RED, 0, 0, alpha);
-      mpg.strokeWeight(OUT_SCALE / 2);
-      mpg.fill(COLOR_RED, 0, 0, alpha / 2);
+      mpg.strokeWeight(OUT_SCALE / 2.0);
+      mpg.fill(COLOR_RED, 0, 0, alpha / 2.0);
 
       mpg.rect(x, y,
-        1.5*(PWIDTH + OUT_SCALE)*mPerlin.noise(x * NOISE_SCALE, y * NOISE_SCALE, PI),
-        1.5*(PWIDTH + OUT_SCALE)*mPerlin.noise(x * NOISE_SCALE, y * NOISE_SCALE, TWO_PI));
+        1.5 * (PWIDTH + OUT_SCALE) * mPerlin.noise(x * NOISE_SCALE, y * NOISE_SCALE, PI),
+        1.5 * (PWIDTH + OUT_SCALE) * mPerlin.noise(x * NOISE_SCALE, y * NOISE_SCALE, TWO_PI));
     }
   }
   mpg.endDraw();
